Move WindowAccumulator to SPI
