+++ core/trino-main/src/test/java/io/trino/operator/TestRealAverageAggregation.java
@@ -17,12 +17,12 @@ import com.google.common.collect.ImmutableList;
 import io.trino.block.BlockAssertions;
 import io.trino.metadata.ResolvedFunction;
 import io.trino.operator.aggregation.AbstractTestAggregationFunction;
-import io.trino.operator.aggregation.WindowAccumulator;
 import io.trino.operator.window.PagesWindowIndex;
 import io.trino.spi.Page;
 import io.trino.spi.block.Block;
 import io.trino.spi.block.BlockBuilder;
 import io.trino.spi.function.AggregationImplementation;
+import io.trino.spi.function.WindowAccumulator;
 import io.trino.spi.function.WindowIndex;
 import io.trino.spi.type.Type;
 import org.junit.jupiter.api.Test;
